refactor(backend): group media storage workflows
