+++ apps/backend/src/mediaAssets/storage/direct.ts
@@ -1,6 +1,6 @@
 import { getMediaAssetsS3Client, getMediaAssetsStorageConfig } from "./config";
 import type { StoreMediaAssetBlobBytesInput } from "./contracts";
-import { storeMediaAssetBlobBytesIfAbsentWithDependencies } from "./promotion";
+import { storeMediaAssetBlobBytesIfAbsentWithDependencies } from "./promotion/promotion";
 
 export function storeMediaAssetBlobBytesIfAbsent(
   input: StoreMediaAssetBlobBytesInput,
